tornado_demo/pyproject.toml
[project]
name = "sapl-tornado-demo"
version = "1.0.0"
description = "Tornado demo application for SAPL Python PEP"
requires-python = ">=3.12"
dependencies = [
    "tornado>=6.4",
    "sapl-tornado>=4.0.0rc1",
    "pyjwt[crypto]>=2.8",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "structlog>=24.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
